Move uri module to root and make Uri public.
